+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import InputContainer from './InputContainer';
-import ListContainer from './ListContainer';
+import InputContainer from './container/InputContainer';
+import ListContainer from './container/ListContainer';
 
 export default function App() {
   return (
