+++ src/store.js
@@ -1,6 +1,6 @@
 import { createStore } from 'redux';
 
-import reducer from '../reducer';
+import reducer from './reducers/reducers';
 
 const store = createStore(reducer);
