+++ src/test/App.test.jsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 
 import { useSelector } from 'react-redux';
 
-import App from './App';
+import App from '../App';
 
 jest.mock('react-redux');
