+++ src/test/presentational/Item.test.jsx
@@ -1,6 +1,6 @@
 import { render, fireEvent } from '@testing-library/react';
 
-import Item from './Item';
+import Item from '../../presentational/Item';
 
 test('Item', () => {
   const task = {
