+++ src/test/presentational/List.test.jsx
@@ -1,6 +1,6 @@
 import { render, fireEvent } from '@testing-library/react';
 
-import List from './List';
+import List from '../../presentational/List';
 
 // test('테스트 #1')
 //
